+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 No unreleased changes
 
+## v0.3.0 (Released 2025-10-06)
+
+* Added `CallerInfo` type and `DefaultCallerInfo` and `GetCallerInfo` functions
+* Enhanced `MarshalJSON` to include wrapped errors
+
 ## v0.2.0 (Released 2025-10-05)
 
 * Added `Wrap` and `Wrapf` functions